requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
tqdm>=4.0.0
requests>=2.0.0
genanki>=0.13.0
